Add isJoined hook
